moved helpers in app.helpers.utils.py | used with statement for subprocess.Popen()
